requirement.txt
fastapi
uvicorn
google-cloud-bigquery
pydantic
requests
